requirements.txt
click>=8.0.0
fastapi[standard]
gitingest
google-generativeai
python-dotenv
slowapi
starlette
tiktoken
uvicorn
